+++ src/settings/LocationLocations/LocationForm/LocationFormContainer.js
@@ -12,6 +12,8 @@ import { useLocationCreate } from '../../../hooks/useLocationCreate';
 import { useLocationUpdate } from '../../../hooks/useLocationUpdate';
 import { LOCATIONS } from '../../../hooks/useLocations';
 import { SERVICE_POINTS } from '../../../hooks/useServicePoints';
+import { REMOTE_STORAGE_CONFIGURATIONS } from '../../../hooks/useRemoteStorageConfigurations';
+import { REMOTE_STORAGE_MAPPINGS } from '../../../hooks/useRemoteStorageMappings';
 
 
 const LocationFormContainer = ({
@@ -89,6 +91,10 @@ const LocationFormContainer = ({
 
     saveData(data)
       .then(onSave)
+      .then(() => {
+        queryClient.invalidateQueries(REMOTE_STORAGE_CONFIGURATIONS);
+        queryClient.invalidateQueries(REMOTE_STORAGE_MAPPINGS);
+      })
       .catch(showSubmitErrorCallout);
   };
 
@@ -103,7 +109,9 @@ const LocationFormContainer = ({
 
 LocationFormContainer.propTypes = {
   onSave: PropTypes.func,
-  servicePointsByName: PropTypes.object,
+  servicePointsByName: PropTypes.objectOf(
+    PropTypes.string,
+  ),
   initialValues: PropTypes.shape({
     id: PropTypes.string,
   }),