+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 10.1.0 (IN PROGRESS)
 
+* [UITEN-327](https://folio-org.atlassian.net/browse/UITEN-327) Invalidate queries after updating location
+
 ## [10.0.0](https://github.com/folio-org/ui-tenant-settings/tree/v10.0.0)(2025-03-12)
 
 * [UITEN-316](https://folio-org.atlassian.net/browse/UITEN-316) Fix code to pass ESLint successfully.